back/pyproject.toml
[project]
name = 'app'
version = '0.0.1'
requires-python = ">=3.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
pythonpath = [ "." ]
testpaths = [ "tests/unit" ]
markers = [ "asyncio" ]
addopts = [
    "--import-mode=importlib",
]

[tool.poetry]
package-mode = false
packages = [
    { include = "app" },
    { include = "tasks" },
    { include = ".api_clients" }
]

[tool.poetry.group.test.dependencies]
pytest = "^9.0.1"
websocket-client = "^1.9.0"
pytest-mock = "^3.15.1"
pytest-asyncio = "^1.3.0"
debugpy = "^1.8.17"
pytest-cov = "^7.0.0"

[tool.poetry.group.tasks.dependencies]
celery = {extras = ["redis"], version = "^5.6.0"}
psycopg2 = "^2.9.11"

[tool.poetry.dependencies]
sqlalchemy = "^2.0.44"
aiohttp = "^3.13.2"
apscheduler = "^3.11.1"
dishka = "^1.7.2"
fastapi = "^0.121.3"
opencv-python = "^4.12.0.88"
passlib = "^1.7.4"
pydantic = "^2.12.4"
pyjwt = "^2.10.1"
pyyaml = "^6.0.3"
requests = "^2.32.5"
perplexityai = "^0.20.0"
pyzbar = "^0.1.9"
aiofiles = "^25.1.0"
python-multipart = "^0.0.20"
aiosqlite = {version = "^0.21.0", extras = ["sqlalchemy"]}
uvicorn = "^0.38.0"
redis = "<6.5.0"
asyncpg = "^0.31.0"
authlib = "^1.6.6"
itsdangerous = "^2.2.0"
pydantic-settings = "^2.12.0"
alembic = "^1.18.0"

[tool.ruff]
target-version = "py312"
src = ["app"]
 
[tool.ruff.lint]
select = [ "E", "F", "I", "RUF" ]
exclude = [ "tests/**" ]
fixable = [ "I", "RUF" ]
ignore = [ "E501" ]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F403", "F401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
